chore: Attempt to wire up emulator connections
