+++ mockzilla-management-ui/common/src/commonMain/kotlin/com/apadmi/mockzilla/desktop/engine/connection/DeviceDetectionUseCase.kt
@@ -1,58 +1,52 @@
 package com.apadmi.mockzilla.desktop.engine.connection
 
 import com.apadmi.mockzilla.desktop.engine.connection.jmdns.ServiceInfoWrapper
-import com.apadmi.mockzilla.lib.models.MetaData
 import com.apadmi.mockzilla.lib.models.MetaData.Companion.parseMetaData
 import com.apadmi.mockzilla.lib.models.RunTarget
-import kotlinx.coroutines.CoroutineScope
-import kotlinx.coroutines.GlobalScope
-import kotlinx.coroutines.channels.BufferOverflow
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.MutableSharedFlow
-import kotlinx.coroutines.flow.MutableStateFlow
-import kotlinx.coroutines.flow.StateFlow
-import kotlinx.coroutines.launch
 import kotlinx.coroutines.sync.Mutex
 import kotlinx.coroutines.sync.withLock
-import org.jetbrains.skia.skottie.Logger
 
-data class DetectedDevice(
-    val connectionName: String,
-    val metaData: MetaData?,
-    val hostAddress: String,
-    val hostAddresses: List<String>,
-    val port: Int,
-    val state: State
-) {
-    enum class State {
-        Resolving,
-        ReadyToConnect,
-        NotYourSimulator,
-        Removed
-    }
-}
 
 interface DeviceDetectionUseCase {
+    suspend fun prepareForConnection(device: DetectedDevice): Result<String>
+
     val onChangeEvent: Flow<Unit>
     val devices: List<DetectedDevice>
 }
 
 class DeviceDetectionUseCaseImpl(
     private val localIpAddress: () -> String,
-    private val adbConnectorUseCase: AdbConnectorUseCase,
-    private val scope: CoroutineScope = GlobalScope
+    private val adbConnectorUseCase: AdbConnectorUseCase
 ) : DeviceDetectionUseCase {
     private val deviceCache = mutableMapOf<String, DetectedDevice>()
+    override suspend fun prepareForConnection(device: DetectedDevice): Result<String> {
+        return when (device.metaData?.runTarget) {
+            RunTarget.AndroidEmulator -> {
+                val adbConnection = device.adbConnection ?: findAdbConnection(device.hostAddresses)
+                ?: return Result.failure(Exception("Failed to detect emulator with adbo"))
+                adbConnectorUseCase.setupPortForwardingIfNeeded(
+                    adbConnection,
+                    0,
+                    device.port
+                ).map { "127.0.0.1:${it.localPort}" }
+            }
+
+            RunTarget.iOSSimulator -> Result.success("127.0.0.1:${device.port}")
+            RunTarget.iOSDevice,
+            RunTarget.AndroidDevice,
+            RunTarget.Jvm,
+            null -> Result.success("${device.hostAddress}:${device.port}")
+        }
+    }
+
     override val onChangeEvent = MutableSharedFlow<Unit>(replay = 1)
 
     private val mutex = Mutex()
     internal suspend fun onChangedServiceEvent(info: ServiceInfoWrapper) = mutex.withLock {
-
-        val adbConnection = adbConnectorUseCase.listConnectedDevices().getOrNull()?.firstOrNull {
-            it.ipAddresses.intersect(info.hostAddresses.toSet()).isNotEmpty()
-        }
+        val adbConnection = findAdbConnection(info.hostAddresses)
         val metaData = runCatching { info.attributes.parseMetaData() }.getOrNull()
-
 
         val state = when (info.state) {
             ServiceInfoWrapper.State.Found -> DetectedDevice.State.Resolving
@@ -84,7 +78,10 @@ class DeviceDetectionUseCaseImpl(
             // callback so ignore this event
             existingDevice?.state == DetectedDevice.State.ReadyToConnect
                     && state == DetectedDevice.State.Resolving -> existingDevice
-            else -> { null }
+
+            else -> {
+                null
+            }
         }
 
         deviceCache[info.connectionName] = (device ?: DetectedDevice(
@@ -93,10 +90,16 @@ class DeviceDetectionUseCaseImpl(
             info.hostAddress,
             info.hostAddresses,
             info.port,
+            adbConnection,
             state
         ))
         onChangeEvent.emit(Unit)
     }
+
+    private suspend fun findAdbConnection(hostAddresses: List<String>) =
+        adbConnectorUseCase.listConnectedDevices().getOrNull()?.firstOrNull {
+            it.ipAddresses.intersect(hostAddresses.toSet()).isNotEmpty()
+        }
 
     override val devices: List<DetectedDevice>
         get() = deviceCache.values.toList()
