+++ samples/demo-kmm/shared/src/commonMain/kotlin/com/apadmi/mockzilla/mock/MockServer.kt
@@ -48,5 +48,6 @@ private val getMyCow = EndpointConfiguration
 
 val mockzillaConfig = MockzillaConfig.Builder()
     .addEndpoint(getMyCow)
+    .setPort(8085)
     .setLogLevel(MockzillaConfig.LogLevel.Verbose)
     .build()
